possible need to convert prio storage to int for proper address in flash

+++ canbus-firewall-avr/src/rules.c
@@ -688,7 +688,7 @@ bool handle_new_rule_data_cmd(Union64 *data, int working_set_index)
         {
             store_new_sequence_number(rules_in_progress.working_sets[working_set_index]);
             rule_t rule_to_save = create_rule_from_working_set(rules_in_progress.working_sets[working_set_index]);            
-            success = save_rule_to_flash(&rule_to_save, &flash_can_ruleset[rule_to_save.prio]);
+            success = save_rule_to_flash(&rule_to_save, &flash_can_ruleset[(int)rule_to_save.prio]);
             
         }
